+++ tests/testthat/test-series-list.R
@@ -32,3 +32,29 @@ patrick::with_parameters_test_that(
     )
   }
 )
+
+test_that("series list$to_struct accepts future-compatible fields", {
+  series <- as_polars_series(list(c(1, 2), c(1, 2, 3)))
+
+  expect_no_warning(series$list$to_struct(c("a", "b")))
+  expect_no_warning(series$list$to_struct(fields = c("a", "b")))
+  expect_no_warning(series$list$to_struct(fields = "max_width"))
+  expect_snapshot(as_polars_df(series$list$to_struct()), cnd_class = TRUE)
+  expect_snapshot(
+    as_polars_df(series$list$to_struct("max_width")),
+    cnd_class = TRUE
+  )
+  expect_snapshot(
+    series$list$to_struct("first_non_null", c("a"), 2),
+    error = TRUE
+  )
+  expect_snapshot(
+    series$list$to_struct(
+      fields = "a",
+      "first_non_null",
+      2
+    ),
+    error = TRUE
+  )
+  expect_snapshot(series$list$to_struct(upper_bound = 2), error = TRUE)
+})